Move duplicated code to NPQ helper
