+++ app/components/finance/ecf/output_payments.rb
@@ -3,7 +3,7 @@
 module Finance
   module ECF
     class OutputPayments < BaseComponent
-      include FinanceHelper
+      include ECFPaymentsHelper
 
       def initialize(output_payments:, breakdown_summary:)
         @output_payments = output_payments
