+++ app/components/finance/npq/payment_overviews/contract_info.rb
@@ -4,7 +4,7 @@ module Finance
   module NPQ
     module PaymentOverviews
       class ContractInfo < BaseComponent
-        include FinanceHelper
+        include NPQPaymentsHelper
 
         def initialize(npq_contracts, npq_lead_provider)
           @npq_contracts = npq_contracts
