+++ app/components/finance/npq/payment_overviews/output_payment_row.html.erb
@@ -1,6 +1,6 @@
 <tr class="gov-table__row">
   <td scope="row" class="govuk-table__cell govuk-body-s"><%= t("finance.output_payment.label") %></td>
   <td scope="row" class="govuk-table__cell govuk-body-s"><%= total_declarations %></td>
-  <td scope="row" class="govuk-table__cell govuk-body-s"><%= number_to_pounds payment_per_participant %></td>
-  <td scope="row" class="govuk-table__cell govuk-body-s"><%= number_to_pounds total %></td>
+  <td scope="row" class="govuk-table__cell govuk-body-s"><%= number_to_pounds output_payment_per_participant %></td>
+  <td scope="row" class="govuk-table__cell govuk-body-s"><%= number_to_pounds output_payment_subtotal %></td>
 </tr>
