+++ app/components/finance/npq/payment_overviews/output_payment_row.rb
@@ -4,7 +4,7 @@ module Finance
   module NPQ
     module PaymentOverviews
       class OutputPaymentRow < BaseComponent
-        include FinanceHelper
+        include NPQPaymentsHelper
 
         attr_accessor :output_payment, :contract, :statement, :npq_lead_provider
 
@@ -13,31 +13,6 @@ module Finance
           @contract = contract
           @statement = statement
           @npq_lead_provider = npq_lead_provider
-        end
-
-        def total
-          output_payment[:subtotal]
-        end
-
-        def payment_per_participant
-          output_payment[:per_participant]
-        end
-
-        def total_declarations
-          if statement.current?
-            ParticipantDeclaration::NPQ
-              .eligible_for_lead_provider(npq_lead_provider)
-              .for_course_identifier(contract.course_identifier)
-              .where(statement_id: nil)
-              .count
-          else
-            statement
-              .participant_declarations
-              .for_course_identifier(contract.course_identifier)
-              .paid_payable_or_eligible
-              .unique_id
-              .count
-          end
         end
       end
     end
