+++ app/components/finance/npq/payment_overviews/payment_overview_total.rb
@@ -4,33 +4,17 @@ module Finance
   module NPQ
     module PaymentOverviews
       class PaymentOverviewTotal < BaseComponent
-        include FinanceHelper
+        include NPQPaymentsHelper
 
-        def initialize(service_fees, output_payments, npq_lead_provider)
+        def initialize(service_fees, output_payment, npq_lead_provider)
           @service_fees = service_fees
-          @output_payments = output_payments
+          @output_payment = output_payment
           @npq_lead_provider = npq_lead_provider
         end
 
       private
 
-        attr_accessor :service_fees, :output_payments, :npq_lead_provider
-
-        def course_total
-          course_payment
-        end
-
-        def course_payment
-          monthly_service_fees + output_payment_subtotal
-        end
-
-        def monthly_service_fees
-          service_fees[:monthly]
-        end
-
-        def output_payment_subtotal
-          output_payments[:subtotal]
-        end
+        attr_accessor :service_fees, :output_payment, :npq_lead_provider
       end
     end
   end
