+++ app/components/finance/npq/payment_overviews/service_fee_row.rb
@@ -4,19 +4,11 @@ module Finance
   module NPQ
     module PaymentOverviews
       class ServiceFeeRow < BaseComponent
-        include FinanceHelper
+        include NPQPaymentsHelper
 
         def initialize(service_fees, contract)
           @service_fees = service_fees
           @contract = contract
-        end
-
-        def total
-          service_fees[:monthly]
-        end
-
-        def payment_per_participant
-          service_fees[:per_participant]
         end
 
         delegate :recruitment_target, to: :contract
