+++ config/locales/en.yml
@@ -221,7 +221,7 @@ en:
       not_yet_known: "Not yet decided"
 
   npq_courses:
-    npq-leading-teaching: Leading Teaching development course breakdown
+    npq-leading-teaching: Leading Teaching course breakdown
     npq-leading-behaviour-culture: Leading Behaviour and Culture course breakdown
     npq-leading-teaching-development: Leading Teacher Development course breakdown
     npq-senior-leadership: Senior Leadership course breakdown
